Re-structure the neural network package
Re-structure the `nn` package to:

* Feature an external half precision module since it is used in
  multiple separate modules.
* Use half precision in the weight JSON format to decrease the disk size
  of the file (88Mb -> 44Mb)
* Move the ffi interface to NVIDIA CUDA into an internal `ffi` module
  that is private to the `nn` package.
* Improve error checking of CUDA.

+++ tools/bootstrap.py
@@ -452,7 +452,7 @@ def dump(args):
                 # cudnn:      [out, in, h, w]
                 value = np.transpose(value, [3, 2, 0, 1])
 
-            serialized = value.flatten().astype(np.float32).tostring()
+            serialized = value.flatten().astype(np.float16).tostring()
             values[var.name] = base64.b85encode(serialized, pad=True).decode('ascii')
 
         json.dump(values, sys.stdout, sort_keys=True)
